+++ packages/cordova-plugin-notificare-push/src/android/NotificarePushPlugin.kt
@@ -1,8 +1,20 @@
 package re.notifica.push.cordova
 
+import android.Manifest
 import android.content.Intent
+import android.content.pm.PackageManager
+import android.net.Uri
+import android.os.Build
 import android.os.Handler
 import android.os.Looper
+import android.provider.Settings
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.annotation.RequiresApi
+import androidx.appcompat.app.AlertDialog
+import androidx.core.app.ActivityCompat
+import androidx.core.app.NotificationManagerCompat
+import androidx.core.content.ContextCompat
 import androidx.lifecycle.Observer
 import org.apache.cordova.CallbackContext
 import org.apache.cordova.CordovaArgs
@@ -11,9 +23,15 @@ import org.apache.cordova.PluginResult
 import org.json.JSONArray
 import org.json.JSONObject
 import re.notifica.Notificare
+import re.notifica.internal.NotificareLogger
 import re.notifica.push.ktx.push
 
 class NotificarePushPlugin : CordovaPlugin() {
+    private var shouldShowRationale = false
+    private var hasOnGoingPermissionRequest = false
+    private var permissionRequestCallback: CallbackContext? = null
+
+    private lateinit var notificationsPermissionLauncher: ActivityResultLauncher<String>
 
     private val allowedUIObserver = Observer<Boolean> { allowedUI ->
         if (allowedUI == null) return@Observer
@@ -30,6 +48,26 @@ class NotificarePushPlugin : CordovaPlugin() {
 
         val intent = cordova.activity.intent
         if (intent != null) onNewIntent(intent)
+
+        notificationsPermissionLauncher = cordova.activity.registerForActivityResult(
+            ActivityResultContracts.RequestPermission()
+        ) { granted ->
+            if (granted) {
+                permissionRequestCallback?.success(PermissionStatus.GRANTED.rawValue)
+            } else {
+                if (!shouldShowRationale &&
+                    !ActivityCompat.shouldShowRequestPermissionRationale(cordova.activity, PUSH_PERMISSION)
+                ) {
+                    permissionRequestCallback?.success(PermissionStatus.PERMANENTLY_DENIED.rawValue)
+                } else {
+                    permissionRequestCallback?.success(PermissionStatus.DENIED.rawValue)
+                }
+            }
+
+            shouldShowRationale = false
+            hasOnGoingPermissionRequest = false
+            permissionRequestCallback = null
+        }
     }
 
     override fun onDestroy() {
@@ -51,6 +89,11 @@ class NotificarePushPlugin : CordovaPlugin() {
             "allowedUI" -> allowedUI(args, callback)
             "enableRemoteNotifications" -> enableRemoteNotifications(args, callback)
             "disableRemoteNotifications" -> disableRemoteNotifications(args, callback)
+            "checkPermissionStatus" -> checkPermissionStatus(args, callback)
+            "shouldShowPermissionRationale" -> shouldShowPermissionRationale(args, callback)
+            "presentPermissionRationale" -> presentPermissionRationale(args, callback)
+            "requestPermission" -> requestPermission(args, callback)
+            "openAppSettings" -> openAppSettings(args, callback)
 
             // Event broker
             "registerListener" -> registerListener(args, callback)
@@ -104,8 +147,143 @@ class NotificarePushPlugin : CordovaPlugin() {
 
     // endregion
 
+    // region Notificare Push
+
+    private fun checkPermissionStatus(@Suppress("UNUSED_PARAMETER") args: CordovaArgs, callback: CallbackContext) {
+        val context = cordova.context ?: run {
+            callback.error("Cannot continue without a context.")
+            return
+        }
+
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+            val granted = NotificationManagerCompat.from(context.applicationContext).areNotificationsEnabled()
+            callback.success(if (granted) PermissionStatus.GRANTED.rawValue else PermissionStatus.PERMANENTLY_DENIED.rawValue)
+            return
+        }
+
+        val granted = ContextCompat.checkSelfPermission(context, PUSH_PERMISSION) == PackageManager.PERMISSION_GRANTED
+        callback.success(if (granted) PermissionStatus.GRANTED.rawValue else PermissionStatus.DENIED.rawValue)
+    }
+
+    private fun shouldShowPermissionRationale(@Suppress("UNUSED_PARAMETER") args: CordovaArgs, callback: CallbackContext) {
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+            callback.success(false)
+            return
+        }
+
+        val activity = cordova.activity ?: run {
+            NotificareLogger.warning("Unable to acquire a reference to the current activity.")
+            callback.error("Unable to acquire a reference to the current activity.")
+            return
+        }
+
+        val result = ActivityCompat.shouldShowRequestPermissionRationale(activity, PUSH_PERMISSION)
+
+        callback.success(result)
+    }
+
+    private fun presentPermissionRationale(@Suppress("UNUSED_PARAMETER") args: CordovaArgs, callback: CallbackContext) {
+        val activity = cordova.activity ?: run {
+            NotificareLogger.warning("Unable to acquire a reference to the current activity.")
+            callback.error("Unable to acquire a reference to the current activity.")
+            return
+        }
+
+        val rationale =
+            if (!args.isNull(0)) {
+                args.getJSONObject(0)
+            } else {
+                callback.error("Missing rationale parameter.")
+                return
+            }
+
+        val title = if (!rationale.isNull("title")) rationale.getString("title") else null
+        val message =
+            if (!rationale.isNull("message")) {
+                rationale.getString("message")
+            } else {
+                callback.error("Missing message parameter.")
+                return
+            }
+        val buttonText =
+            if (!rationale.isNull("buttonText")) rationale.getString("buttonText")
+            else activity.getString(android.R.string.ok)
+
+        try {
+            NotificareLogger.debug("Presenting permission rationale for notifications.")
+
+            activity.runOnUiThread {
+                AlertDialog.Builder(activity)
+                    .setTitle(title)
+                    .setMessage(message)
+                    .setCancelable(false)
+                    .setPositiveButton(buttonText, null)
+                    .setOnDismissListener { callback.success() }
+                    .show()
+            }
+        } catch (e: Exception) {
+            callback.error("Unable to present the rationale alert.")
+        }
+    }
+
+    private fun requestPermission(@Suppress("UNUSED_PARAMETER") args: CordovaArgs, callback: CallbackContext) {
+        val activity = cordova.activity ?: run {
+            callback.error("Cannot continue without a activity.")
+            return
+        }
+
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+            val granted = NotificationManagerCompat.from(activity.applicationContext).areNotificationsEnabled()
+            callback.success(if (granted) PermissionStatus.GRANTED.rawValue else PermissionStatus.PERMANENTLY_DENIED.rawValue)
+            return
+        }
+
+        if (hasOnGoingPermissionRequest) {
+            NotificareLogger.warning("A request for permissions is already running, please wait for it to finish before doing another request.")
+            callback.error("A request for permissions is already running, please wait for it to finish before doing another request.")
+            return
+        }
+
+        val granted = ContextCompat.checkSelfPermission(activity, PUSH_PERMISSION) == PackageManager.PERMISSION_GRANTED
+
+        if (granted) {
+            callback.success(PermissionStatus.GRANTED.rawValue)
+            return
+        }
+
+        shouldShowRationale = ActivityCompat.shouldShowRequestPermissionRationale(activity, PUSH_PERMISSION)
+        hasOnGoingPermissionRequest = true
+        permissionRequestCallback = callback
+
+        notificationsPermissionLauncher.launch(PUSH_PERMISSION)
+    }
+
+    private fun openAppSettings(@Suppress("UNUSED_PARAMETER") args: CordovaArgs, callback: CallbackContext) {
+        try {
+            val context = cordova.context ?: run {
+                callback.error("Cannot continue without a context.")
+                return
+            }
+
+            val packageName = Uri.fromParts("package", context.packageName, null)
+            context.startActivity(
+                Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS, packageName)
+                    .addCategory(Intent.CATEGORY_DEFAULT)
+                    .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+                    .addFlags(Intent.FLAG_ACTIVITY_NO_HISTORY)
+                    .addFlags(Intent.FLAG_ACTIVITY_EXCLUDE_FROM_RECENTS)
+            )
+
+            callback.void()
+        } catch (e: Exception) {
+            callback.error("Unable to open the app settings.")
+        }
+    }
+
+    // endregion
+
     private fun registerListener(@Suppress("UNUSED_PARAMETER") args: CordovaArgs, callback: CallbackContext) {
-        NotificarePushPluginEventBroker.setup(object : NotificarePushPluginEventBroker.Consumer {
+        NotificarePushPluginEventBroker.setup(preferences, object : NotificarePushPluginEventBroker.Consumer {
             override fun onEvent(event: NotificarePushPluginEventBroker.Event) {
                 val payload = JSONObject()
                 payload.put("name", event.name)
@@ -127,6 +305,34 @@ class NotificarePushPlugin : CordovaPlugin() {
                 callback.sendPluginResult(result)
             }
         })
+    }
+
+    private companion object {
+        @RequiresApi(Build.VERSION_CODES.TIRAMISU)
+        private const val PUSH_PERMISSION = Manifest.permission.POST_NOTIFICATIONS
+    }
+
+    internal enum class PermissionStatus {
+        DENIED,
+        GRANTED,
+        PERMANENTLY_DENIED;
+
+        internal val rawValue: String
+            get() = when (this) {
+                DENIED -> "denied"
+                GRANTED -> "granted"
+                PERMANENTLY_DENIED -> "permanently_denied"
+            }
+
+        internal companion object {
+            internal fun parse(status: String): PermissionStatus? {
+                values().forEach {
+                    if (it.rawValue == status) return it
+                }
+
+                return null
+            }
+        }
     }
 }