+++ packages/cordova-plugin-notificare-push/src/js/index.ts
@@ -1,7 +1,10 @@
 import { NotificarePush } from './notificare-push';
 import { bootstrap } from './events';
 
+export * from './models/notificare-notification-delivery-mechanism';
 export * from './models/notificare-system-notification';
+
+export * from './permissions';
 
 export default NotificarePush;