+++ packages/cordova-plugin-notificare-user-inbox/src/js/index.ts
@@ -1,0 +1,6 @@
+import { NotificareUserInbox } from './notificare-user-inbox';
+
+export * from './models/notificare-user-inbox-item';
+export * from './models/notificare-user-inbox-response';
+
+export default NotificareUserInbox;